pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arxiv>=2.3.1",
    "chromadb>=1.3.5",
    "faiss-cpu>=1.13.0",
    "fastembed>=0.7.3",
    "flashrank>=0.2.10",
    "google-generativeai>=0.8.3",
    "ipykernel>=7.1.0",
    "langchain>=0.3.0",
    "langchain-chroma>=0.1.0",
    "langchain-community>=0.3.0",
    "langchain-google-genai>=2.0.5",
    "langchain-groq>=0.2.0",
    "langchain-mistralai>=0.2.0",
    "langgraph>=0.2.0",
    "pillow>=11.3.0",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "rapidocr-onnxruntime>=1.4.4",
    "structlog>=25.5.0",
    "tiktoken>=0.12.0",
]
